Move ToyMCDataset to pfmatch.io.dataset
